+++ libs/base/ui/graphical/icon/src/icon.types.ts
@@ -1,0 +1,79 @@
+export enum IconTypes {
+  Add = 'add',
+  Cart = 'cart',
+  Close = 'close',
+  Collapse = 'collapse',
+  DropUp = 'drop-up',
+  Dropdown = 'dropdown',
+  Expand = 'expand',
+  Image = 'image',
+  Decrease = 'decrease',
+  Increase = 'increase',
+  Info = 'info',
+  InputError = 'inputError',
+  Mark = 'mark',
+  Minus = 'minus',
+  Remove = 'remove',
+  Report = 'report',
+  Rocket = 'rocket',
+  Search = 'search',
+  Trash = 'trash',
+  Visible = 'visible',
+  Invisible = 'invisible',
+  Wishlist = 'wishlist',
+  Customer = 'customer',
+  Ratings = 'ratings',
+  Sales = 'sales',
+  User = 'user',
+  Users = 'users',
+  ModeDark = 'mode-dark',
+  ModeLight = 'mode-light',
+  Orders = 'orders',
+  Parcel = 'parcel',
+  Printer = 'printer',
+  Products = 'products',
+  Profile = 'profile',
+  Desktop = 'desktop',
+  Mobile = 'mobile',
+  Tablet = 'tablet',
+  Error = 'error',
+  Success = 'success',
+  Warning = 'warning',
+  Planet = 'planet',
+  Actions = 'actions',
+  ActionsHorizontal = 'actionsHorizontal',
+  Back = 'back',
+  Bottom = 'bottom',
+  Calendar = 'calendar',
+  Copy = 'copy',
+  Create = 'create',
+  Disconnect = 'disconnect',
+  Drag = 'drag',
+  DragSmall = 'dragSmall',
+  Edit = 'edit',
+  FastForward = 'fastForward',
+  File = 'file',
+  Filters = 'filters',
+  Front = 'front',
+  Help = 'help',
+  Imports = 'imports',
+  InputStepper = 'inputStepper',
+  Link = 'link',
+  Loader = 'loader',
+  Locker = 'locker',
+  Maximize = 'maximize',
+  Menu = 'menu',
+  Minimize = 'minimize',
+  NavigationArrow = 'navigationArrow',
+  Reset = 'reset',
+  Settings = 'settings',
+  Sort = 'sort',
+  Star = 'star',
+  Top = 'top',
+  TextAreaResizable = 'textAreaResizable',
+  Filter = 'filter',
+  Refresh = 'refresh',
+  Carrier = 'carrier',
+  Location = 'location',
+  Login = 'login',
+}
