+++ libs/base/ui/graphical/icon/src/index.ts
@@ -2,3 +2,4 @@ export * from './component';
 export * from './icon.component';
 export * from './icon.factory';
 export * from './icon.model';
+export * from './icon.types';
