+++ libs/template/themes/icons/src/icon.model.ts
@@ -1,105 +1,33 @@
-export enum IconTypes {
-  Add = 'add',
-  CartAdd = 'cart-add',
-  Cart = 'cart',
-  Close = 'close',
-  Collapse = 'collapse',
-  DropUp = 'drop-up',
-  Dropdown = 'dropdown',
-  Expand = 'expand',
-  Image = 'image',
-  Decrease = 'decrease',
-  Increase = 'increase',
-  Info = 'info',
-  InputError = 'inputError',
-  Mark = 'mark',
-  Minus = 'minus',
-  Remove = 'remove',
-  Report = 'report',
-  Rocket = 'rocket',
-  Search = 'search',
-  Trash = 'trash',
-  Visible = 'visible',
-  Invisible = 'invisible',
-  Wishlist = 'wishlist',
-  Administration = 'administration',
-  Catalog = 'catalog',
-  Content = 'content',
-  Customer = 'customer',
-  Dashboard = 'dashboard',
-  Maintenance = 'maintenance',
-  Marketplace = 'marketplace',
-  Merchandising = 'merchandising',
-  MerchantProfile = 'merchantProfile',
-  Ratings = 'ratings',
-  Sales = 'sales',
-  User = 'user',
-  Users = 'users',
-  Dashboards = 'dashboards',
-  ModeDark = 'mode-dark',
-  ModeLight = 'mode-light',
-  Merchant = 'merchant',
-  Orders = 'orders',
-  Parcel = 'parcel',
-  Printer = 'printer',
-  Products = 'products',
-  Profile = 'profile',
-  Desktop = 'desktop',
-  Mobile = 'mobile',
-  Tablet = 'tablet',
-  Error = 'error',
-  Success = 'success',
-  Warning = 'warning',
-  America = 'america',
-  France = 'france',
-  Germany = 'germany',
-  Planet = 'planet',
-  Spain = 'spain',
-  Actions = 'actions',
-  ActionsHorizontal = 'actionsHorizontal',
-  Back = 'back',
-  Bottom = 'bottom',
-  Calendar = 'calendar',
-  Copy = 'copy',
-  Create = 'create',
-  Disconnect = 'disconnect',
-  Drag = 'drag',
-  DragSmall = 'dragSmall',
-  Edit = 'edit',
-  FastForward = 'fastForward',
-  File = 'file',
-  Filters = 'filters',
-  Front = 'front',
-  Help = 'help',
-  Imports = 'imports',
-  InputStepper = 'inputStepper',
-  Integration = 'integration',
-  Link = 'link',
-  Loader = 'loader',
-  Locker = 'locker',
-  Maximize = 'maximize',
-  Menu = 'menu',
-  Minimize = 'minimize',
-  NavigationArrow = 'navigationArrow',
-  Placeholder = 'placeholder',
-  Reset = 'reset',
-  Settings = 'settings',
-  Sort = 'sort',
-  Star = 'star',
-  TextAreaResizable = 'textAreaResizable',
-  Top = 'top',
-  Filter = 'filter',
-  GoldPartner = 'goldPartner',
-  Popular = 'popular',
-  SilverPartner = 'silverPartner',
-  BackArrow = 'backArrow',
-  Refresh = 'refresh',
-  CheckMark = 'checkMark',
-  FreezerCategory = 'freezerCategory',
-  FreshCategory = 'freshCategory',
-  ChillerCategory = 'chillerCategory',
-  AmbientCategory = 'ambientCategory',
-  Carrier = 'carrier',
-  Location = 'location',
-  Login = 'login',
-}
+import { IconTypes as UiIconTypes } from '@spryker-oryx/ui/icon';
+
+export const IconTypes = {
+  ...UiIconTypes,
+  CartAdd: 'cart-add',
+  Administration: 'administration',
+  Catalog: 'catalog',
+  Content: 'content',
+  Dashboard: 'dashboard',
+  Maintenance: 'maintenance',
+  Marketplace: 'marketplace',
+  Merchandising: 'merchandising',
+  MerchantProfile: 'merchantProfile',
+  Dashboards: 'dashboards',
+  Merchant: 'merchant',
+  America: 'america',
+  France: 'france',
+  Germany: 'germany',
+  Spain: 'spain',
+  Integration: 'integration',
+  Placeholder: 'placeholder',
+  GoldPartner: 'goldPartner',
+  Popular: 'popular',
+  SilverPartner: 'silverPartner',
+  // TODO: check if we can use `back`
+  BackArrow: 'backArrow',
+  // TODO: check if we can use `mark`
+  CheckMark: 'checkMark',
+  FreezerCategory: 'freezerCategory',
+  FreshCategory: 'freshCategory',
+  ChillerCategory: 'chillerCategory',
+  AmbientCategory: 'ambientCategory',
+};
